Set top matches with GT as labeled in RCNN
In some cases we may end up with zero valid proposals (no proposal with IoU over 0.5).
We use the same technique as in RPN target were we argmax the GT and labeled them as such.
Now we end up with at least one valid RCNN regression target for each GT.