Add other tests for RCNNTarget. Resolves #29
Added a test for the empty case (all proposals are background or ignored), in which case we should set the 'best' proposal as foreground.
Also added a test for multiple overlap, where we have several foreground proposals for a single gt box, and we should be setting only the best one as foreground and ignore the rest. Luminoth is currently failing this one.