Improve tests for RCNNTarget.
We're changing the code in `RCNNTarget._build` to pure tensorflow, which doesn't cast `int`s to `float`s implicitly as much as numpy does, so we're setting our tf constants as dtype=tf.float32 explicitly.

We're also adding a new test for RCNNTarget that checks we're getting the right behaviour with multiple gt_boxes.