Make classes.json file created by 'transform' consider merged classes
Now when there are merged classes the class.json file, which is used for
labeling predicted videos/pics, will be empty. This fixes a problem in
which all predictions generated from a dataset that had merged classes
would be labeled with the label of the first class.

Eg.: If you generated a dataset based on coco that filtered all classes
except car, bus, and truck, and that also merged these classes into a
single class. The predictions generated by predict.py or the web server
would be labeled as car, when in fact they could've been a truck or a
bus too.

The idea behind completely removing the label instead of picking a new
one is that a model that was trained to predict a single type of object,
would portray this information more globally, instead of on an per
object basis. For example in the name of the .jpg or .mpg file it
created, or something like this.

Still, there are probably use cases in which it would be useful to let
the user pick the label. This could be added in the future after
choosing a good console argument name for this parameter, and seeing if
we could somehow merge it with the --merge argument while also
maintaining the ability to have the label be empty.