Initial implementation of checkpoint-related commands
Provides basic support for checkpoint management with the following
commands: create, delete, export, import, info and list.

A checkpoint index is stored in `~/.luminoth/`, keeping track of all
the checkpoints available in the system (which can be inspected by way
of the `list` and `info` command).

Checkpoints may be created with the `create` command, which collects
all the required file into a folder inside `~/.luminoth/` or removed
with the `delete` commmand. They can also be shared with the `import`
and `export` commands, which pack the files into a tar file.

The feature still needs more work, mainly in making it more robust.