Fix SSD proposal only returning the best prediction for each feature
Before this commit, we used to chose the prediction with the highest
class probability for each feature in SSD's feature map.
This would filter a lot of the predictions that weren't the best for
their respective feature.

Now we just chose the best matches regardless of from which feature
they may originate from.

Therefore we now can get all the predictions with probability > 0.01
if we wish to. This is useful for calculating the mAP, which requires
us to send the top-k predictions.