Fix base_network checkpoint loader creating constants for all vars (SSD)
Previously we used tf.assign to load the variables in the checkpoint
file into the variables in the base network (eg.: vgg, resnet), this had
the unintended side effect of creating a constant tensor for each
variable + tf.assign combo. This was bad cause it made our .meta file be
enormous (as big as the checkpoint file) and our model building phase
take much longer than needed.

Implemented only in SSD in this particular commit, but will be expanded
to all other models in a following commit.

Also improved some scope and operation names in SSD.