Add detection datasets, from torchvision's reference detection datasets
These scripts are taken from torchvision's detection datasets
references.

We are going to be adding native support for training on cloud services,
so we need to take control of the whole dataloading pipeline.

While this is not strictly necesary at this point, we want to be able to
eventually update our torchvision version without destroying our data
loading pipeline.